move concrete operation types into request.rs
